Clown Car is hidden from the uplink and now admin spawn only
Why the Hell Shadow left this unchanged is beyond me, but here you go. Clown car is busted as hell, and barely counterable unless you're real damn lucky. If someone wants to fix it, go for it, but I doubt anyone wants it to be more than just admin spawn only.